gloas: address §6 envelope/proposer review findings
- BlindedExecutionPayloadEnvelope.BuilderIndex: use the named gloas.BuilderIndex
  (as the bid does; SSZ-identical). The per-file sszgen directive can't resolve a
  sibling-file type, so switch it to package mode (-path .) like beacon_block.go.
- Note both ExecutionRequests fields track the pinned pre-EIP-8282 Gloas spec; a
  node-side variant follows when the target devnet adopts EIP-8282 (builder
  deposit/exit requests).
- submitGloasProposal: the self-build builder now triggers the §6 envelope even
  when its own block submission fails — the envelope is a cluster round the other
  operators join regardless, so the builder shouldn't abstain.
- startEnvelopeDuty: drop the post-consensus context it forwarded (cancelled when
  the block duty ends); document that the wired starter must dispatch async with a
  node-scoped context.

+++ protocol/v2/types/gloas/execution_payload_envelope_test.go
@@ -13,7 +13,7 @@ func TestBlindedExecutionPayloadEnvelopeRoundTrip(t *testing.T) {
 	in := &BlindedExecutionPayloadEnvelope{
 		PayloadRoot:           phase0.Root{0x01},
 		ExecutionRequests:     &electra.ExecutionRequests{},
-		BuilderIndex:          uint64(BuilderIndexSelfBuild),
+		BuilderIndex:          BuilderIndexSelfBuild,
 		BeaconBlockRoot:       phase0.Root{0x02},
 		ParentBeaconBlockRoot: phase0.Root{0x03},
 	}
@@ -23,7 +23,7 @@ func TestBlindedExecutionPayloadEnvelopeRoundTrip(t *testing.T) {
 	out := &BlindedExecutionPayloadEnvelope{}
 	require.NoError(t, out.UnmarshalSSZ(b))
 	require.Equal(t, in.PayloadRoot, out.PayloadRoot)
-	require.Equal(t, uint64(BuilderIndexSelfBuild), out.BuilderIndex)
+	require.Equal(t, BuilderIndexSelfBuild, out.BuilderIndex)
 	require.Equal(t, in.BeaconBlockRoot, out.BeaconBlockRoot)
 	require.Equal(t, in.ParentBeaconBlockRoot, out.ParentBeaconBlockRoot)
